Splitting of transcription and moving of some kaldi transcription to common
