+++ elpis/engines/common/objects/transcription.py
@@ -2,7 +2,7 @@ from abc import abstractmethod
 from pathlib import Path
 from elpis.engines.common.objects.model import Model
 from elpis.engines.common.objects.fsobject import FSObject
-
+from typing import Callable, Dict, Tuple
 
 class Transcription(FSObject):
     _config_file = "transcription.json"
@@ -16,6 +16,7 @@ class Transcription(FSObject):
         self._exporter = None
         self.config['exporter'] = None
         self.config['has_been_transcribed'] = False
+        self.config['stage_status'] = {}
 
     @classmethod
     def load(cls, base_path: Path):
@@ -39,6 +40,18 @@ class Transcription(FSObject):
     @status.setter
     def status(self, value: str):
         self.config['status'] = value
+
+    @property
+    def stage_status(self):
+        return self.config['stage_status']
+
+    @stage_status.setter
+    def stage_status(self, vals: Tuple[str, str, str]):
+        stage, status, message = vals
+        stage_status = self.config['stage_status']
+        stage_status[stage]['status'] = status
+        stage_status[stage]['message'] = message
+        self.config['stage_status'] = stage_status
     
     @property
     def state(self):
@@ -54,10 +67,16 @@ class Transcription(FSObject):
     @property
     def has_been_transcribed(self):
         return self.config['has_been_transcribed']
-    
+
     @property
     def exporter(self):
         return self._exporter
+
+    def build_stage_status(self, stage_names: Dict[str, str]):
+        for stage_file, stage_name in stage_names.items():
+            stage_status = self.config['stage_status']
+            stage_status.update({stage_file: {'name': stage_name, 'status': 'ready', 'message': ''}})
+            self.config['stage_status'] = stage_status
 
     @abstractmethod
     def transcribe(self, *args, **kwargs):
