+++ elpis/engines/kaldi/inference/gmm-decode/0_feature_vec.sh
@@ -5,6 +5,8 @@
 # Put typical Kaldi binaries onto the path
 . ./path.sh
 
+export PATH=$PATH:/kaldi/src/online2bin
+
 # Extract feature vectors for online training
 echo "==== Extracting Feature Vectors ===="
 steps/make_mfcc.sh --nj 1 data/infer exp/make_mfcc/infer mfcc
