+++ elpis/engines/kaldi/inference/gmm-decode/3_transcription_best_path.sh
@@ -5,6 +5,8 @@
 # Put typical Kaldi binaries onto the path
 . ./path.sh
 
+export PATH=$PATH:/kaldi/src/online2bin
+
 echo "==== Finding Best Path (Transcription) ===="
 lattice-1best \
     ark:exp/tri1_online/decode/lattices.ark \
