+++ elpis/engines/kaldi/inference/gmm-decode/4_word_boundaries.sh
@@ -5,6 +5,8 @@
 # Put typical Kaldi binaries onto the path
 . ./path.sh
 
+export PATH=$PATH:/kaldi/src/online2bin
+
 echo "==== Adding Word Boundaries to FST ===="
 lattice-align-words \
     data/lang/phones/word_boundary.int \
