+++ elpis/engines/kaldi/inference/gmm-decode/5_lattice_to_ctm.sh
@@ -5,6 +5,8 @@
 # Put typical Kaldi binaries onto the path
 . ./path.sh
 
+export PATH=$PATH:/kaldi/src/online2bin
+
 echo "==== Converting Lattice to CTM Format ===="
 nbest-to-ctm \
     ark,t:data/infer/1best-fst-word-aligned.tra \
